handle caprica errors better

+++ pyro/ProcessManager.py
@@ -145,7 +145,9 @@ class ProcessManager:
         :return: ProcessState (SUCCESS, FAILURE, INTERRUPTED, ERRORS)
         """
         command_size = len(command)
-
+        
+        using_caprica = command.lower().find("caprica.exe") != -1
+        
         if command_size > 32766:
             ProcessManager.log.error(f'Cannot create process because command exceeds max length: {command_size}')
             return ProcessState.FAILURE
@@ -170,7 +172,20 @@ class ProcessManager:
             'Starting'
         )
 
-        line_error = re.compile(r'(.*)(\(-?\d*\.?\d+,-?\d*\.?\d+\)):\s+(.*)')
+        # PapyrusCompiler's error structure:
+        # filename      [example.psc]   (.*)
+        # location      [(69,420)]      (\(-?\d*\.?\d+,-?\d*\.?\d+\))
+        # error message [: bad molly]   :\s+(.*)
+        line_error_papyrus = re.compile(r'(.*)(\(-?\d*\.?\d+,-?\d*\.?\d+\)):\s+(.*)')
+        
+        # Caprica's error structure:
+        # filename      [example.psc ]  (.*)\s+
+        # location      [(69, 4:20)]    (\(-?\d*\.?\d+, -?\d*\.?\d+:-?\d*\.?\d+\))
+        # error/warning [: Warning 1: ] :\s+(.*):\s+
+        # error message [bad molly]     (.*)
+        line_error_caprica = re.compile(r'(.*)\s+(\(-?\d*\.?\d+, -?\d*\.?\d+:-?\d*\.?\d+\)):\s+(.*):\s+(.*)')
+        
+        line_error = line_error_caprica if using_caprica else line_error_papyrus
 
         error_count = 0
 
@@ -179,14 +194,28 @@ class ProcessManager:
                 if (line := process.stdout.readline().strip()) != '':
                     if startswith(line, exclusions):
                         continue
-
+                    
                     if (match := line_error.search(line)) is not None:
-                        path, location, message = match.groups()
+                        if (using_caprica):
+                            path, location, message_type, message = match.groups()
+                        else:
+                            path, location, message = match.groups()
                         head, tail = os.path.split(path)
-                        ProcessManager.log.error(f'COMPILATION FAILED: '
-                                                 f'{os.path.basename(head)}\\{tail}{location}: {message}')
-                        process.terminate()
-                        error_count += 1
+                        if not head:
+                            script_path = tail
+                        else:
+                            script_path = f'{os.path.basename(head)}\\{tail}'
+
+                        pyro_message = f'{script_path}{location}: {message}'
+                        if (not using_caprica or message_type.lower() == "fatal error"):
+                            ProcessManager.log.error(f'COMPILATION FAILED: {pyro_message}')
+                            process.terminate()
+                            error_count += 1
+                        elif (message_type.lower() == "error"):
+                            ProcessManager.log.error(f'COMPILATION ERROR: {pyro_message}')
+                            error_count += 1
+                        else:
+                            ProcessManager.log.warning(f'{message_type}: {pyro_message}')
 
                     elif startswith(line, ('Error', 'Fatal Error'), ignorecase=True):
                         ProcessManager.log.error(line)
